Fixes issue where the Reply-To field keeps getting appended to if you have multiple email recipients
When there are multiple recipients and each recipient has a Reply-To
email address (either explicitly specified or taken from email form
field), as the script iterates through the list of recipients to send
emails, the Reply-To addresses get appended for each successive
recipient.

I.e. if there are 4 recipients and each have a Reply-To address, the
fourth recipient will have all four recipients' Reply-To addresses in
the field.

This is because the Email::replyTo calls Email::addCustomHeader which
appends the text rather than set it.

So the solution was to create a setReplyTo function (as opposed to
'append' replyTo.