FIX: Ensured that the data-add-url single quotes are not processed by join_links
Due to the fact that the last bracket meant that join_links was actually sent URL + ' (eg; http://myurl.com' ), this caused problems when the quote was getting escaped - and actually prevented users from adding form fields in the CMS, with a 500. This fix ensures the closing quote is not processed by join_links.