MINOR: fix for styling for one or more buttons that don't need the list styling for the step navigation. Will add the action nav which just shows the buttons
MINOR: fix for styling for one or more buttons that don't need the list styling for the step navigation. Will add the action nav which just shows the buttons

MINOR: fix for styling for one or more buttons that don't need the list styling for the step navigation. Will add the action nav which just shows the buttons